Refactor external manager support
Attempted to remove `--bind-to` entirely and discovered that the port
number does not need to be fixed. It turns out that the issue I
originally ran into while experimenting with this is that the Julia was
only listening to the external interface while `kubectl port-forward`
tries is forwarding from the pod's localhost interface.